fix(integration-ai): prevent memory leaks in lazy-loaded modules
